server finalized

+++ util/utils.py
@@ -41,12 +41,15 @@ def read_json_file(file_path: str):
 
 
 def createResume(filename: str, isSilent: bool = True, texliveonfly=True):
+    # print("Executing for ", filename)
     os.chdir(os.path.join(baseDir, builderDirName))
+    # get only first part of the filename
+    name = filename.split('.')[0]
     try:
-        isSuccess, discription = creatRumeFromSystem(
-            texliveonfly)  # type: ignore
+        isSuccess, discription = creatResumeFromSystem(
+            name, texliveonfly)  # type: ignore
 
-        app.logger.info(str(discription))
+        # app.logger.info(str(discription))
 
         if not isSuccess:
             raise Exception(discription)
@@ -61,42 +64,47 @@ def createResume(filename: str, isSilent: bool = True, texliveonfly=True):
     allfiles.remove('resumecustom.cls')
     allfiles.remove('texliveonfly.py')
 
-    # allfiles.remove('test.py')
-
+    # app.logger.info({"allfiles": allfiles})
+    
     try:
-        allfiles.remove('resume.pdf')
+        allfiles.remove(f'{name}.pdf')
     except ValueError as e:
         raise Exception('🚫 Error in creating the resume', e, allfiles)
 
     # allfiles.remove('resume.tex') # for testing only
     for i in allfiles:
-        os.remove(os.path.join(baseDir, builderDirName, i))
+        if name in i:
+            os.remove(os.path.join(baseDir, builderDirName, i))
 
     # os.chdir(baseDir)
     # print(os.path.join(baseDir, builderDirName, 'resume.pdf'),os.path.join(outputDir, filename))
-    shutil.move(os.path.join(baseDir, builderDirName, 'resume.pdf'),
+    shutil.move(os.path.join(baseDir, builderDirName, f'{name}.pdf'),
                 os.path.join(outputDir, filename))
 
     return os.path.join(str(outputDir), filename)  # type: ignore
 
 
-def creatRumeFromSystem(texliveonfly=True):
+def creatResumeFromSystem(name, texliveonfly=True):
     "error  binary not found showing in the exit code 1"
     # print('testing')
 
     # get the path of the pdflatex
     pdflatexPath = textlivePath + "/pdflatex"
     texliveonflyFilePath = os.path.join(buildDir, "texliveonfly.py")
-    textliveonflyCommand = ''
+
+    toGenerateResumeTexFile = os.path.join(buildDir, f"{name}.tex")
+    pdflatexCommand = f'{pdflatexPath} {toGenerateResumeTexFile}'
+
     if texliveonfly:
-        textliveonflyCommand = f'python3 {texliveonflyFilePath} --texlive_bin={textlivePath} -c '
-    
-    toGenerateResumeTexFile = os.path.join(buildDir, "resume.tex")
+        textliveonflyCommand = f'python3 {texliveonflyFilePath} --texlive_bin={textlivePath} '
+        args = f'-a "-synctex=1 -interaction=nonstopmode -jobname={name}" -c '
+        command = f'{textliveonflyCommand} {args} {pdflatexCommand}'
+    else:
+        command = pdflatexCommand
 
-    command = f'{textliveonflyCommand} {pdflatexPath} {toGenerateResumeTexFile}'
-
-    # app.logger.info({"command": command})
+    app.logger.info({"command": command})
     # generate the resume itself
+    # print(command)
     try:
         (success, error), output = runSystemCommad(command)
         # app.logger.info({"success": success, "error": error})
@@ -106,7 +114,7 @@ def creatRumeFromSystem(texliveonfly=True):
             app.logger.error("error", error.decode())
             raise Exception(error.decode())
         elif success and 'in house texliveonfly' in success.decode():
-            print("success")
+            print("success in generating the resume")
             # app.logger.info("success", success.decode())
             return True, (success)
             # print("Output: ", success.decode())
